drf/requirements/main.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/main.in
#
asgiref==3.3.1
    # via django
django==3.1.7
    # via djangorestframework
djangorestframework==3.12.4
    # via -r requirements/main.in
itsdangerous==2.0.1
    # via -r requirements/main.in
python-dotenv==0.18.0
    # via -r requirements/main.in
pytz==2021.1
    # via django
sqlparse==0.4.1
    # via django
